+++ src/Manager/TagManager.php
@@ -10,6 +10,7 @@ namespace HeimrichHannot\HeadBundle\Manager;
 
 use Contao\Controller;
 use Contao\StringUtil;
+use Contao\System;
 use HeimrichHannot\HeadBundle\Head\TagInterface;
 
 class TagManager