+++ src/Tag/Meta/MetaDescription.php
@@ -8,6 +8,8 @@
 
 namespace HeimrichHannot\HeadBundle\Tag\Meta;
 
+use Contao\Controller;
+use Contao\StringUtil;
 use HeimrichHannot\HeadBundle\Head\AbstractMetaTag;
 
 class MetaDescription extends AbstractMetaTag
@@ -18,4 +20,20 @@ class MetaDescription extends AbstractMetaTag
      * @var string
      */
     protected static $name = 'description';
+
+    /**
+     * @inheritdoc
+     */
+    public function generate()
+    {
+        $content = parent::getContent();
+
+        $content = StringUtil::decodeEntities($content);
+        $content = Controller::replaceInsertTags($content, false);
+        $content = strip_tags($content);
+        $content = str_replace("\n", ' ', $content);
+        $content = \StringUtil::substr($content, 320);
+
+        return sprintf('<%s %s="%s" content="%s">', static::$tag, static::$key, static::$name, $content);
+    }
 }