+++ src/Tag/Misc/Title.php
@@ -19,6 +19,6 @@ class Title extends AbstractTag
      */
     public function generate()
     {
-        return sprintf('<title>%s</title>', $this->getContent());
+        return sprintf('<title>%s</title>', htmlentities($this->getContent(), ENT_COMPAT, \Config::get('characterSet')));
     }
 }